requirements.txt
pymupdf>=1.24.0
chromadb>=0.5.0
ollama>=0.3.0
click>=8.1.0
numpy>=1.26.0
requests>=2.31.0
tqdm>=4.66.0
python-dotenv>=1.0.0
